service/scheduler: schedule, add timezone
- Add Location type
- Add gocql type checks
- Move all Schedule related types to model
- Auto tasks use local timezone
- Scheduler add Location
- Cron add UT to ensure the proper way of handling timezones
