+++ pkg/cmd/scylla-manager/tasks.go
@@ -8,11 +8,23 @@ import (
 
 	"github.com/scylladb/scylla-manager/pkg/service/healthcheck"
 	"github.com/scylladb/scylla-manager/pkg/service/scheduler"
+	"github.com/scylladb/scylla-manager/pkg/util/timeutc"
 	"github.com/scylladb/scylla-manager/pkg/util/uuid"
 )
 
 func healthCheckModeProperties(mode healthcheck.Mode) json.RawMessage {
 	return json.RawMessage(`{"mode": "` + mode.String() + `"}`)
+}
+
+func localTimezone() scheduler.Timezone {
+	if timeutc.LocalName != "" {
+		v, err := time.LoadLocation(timeutc.LocalName)
+		if err == nil {
+			return scheduler.NewTimezone(v)
+		}
+	}
+
+	return scheduler.NewTimezone(nil)
 }
 
 func makeAutoHealthCheckTasks(clusterID uuid.UUID) []*scheduler.Task {
@@ -23,7 +35,8 @@ func makeAutoHealthCheckTasks(clusterID uuid.UUID) []*scheduler.Task {
 			Enabled:   true,
 			Name:      "cql",
 			Sched: scheduler.Schedule{
-				Cron: scheduler.NewCronEvery(15 * time.Second),
+				Cron:     scheduler.NewCronEvery(15 * time.Second),
+				Timezone: localTimezone(),
 			},
 			Properties: healthCheckModeProperties(healthcheck.CQLMode),
 		},
@@ -33,7 +46,8 @@ func makeAutoHealthCheckTasks(clusterID uuid.UUID) []*scheduler.Task {
 			Enabled:   true,
 			Name:      "rest",
 			Sched: scheduler.Schedule{
-				Cron: scheduler.NewCronEvery(1 * time.Minute),
+				Cron:     scheduler.NewCronEvery(1 * time.Minute),
+				Timezone: localTimezone(),
 			},
 			Properties: healthCheckModeProperties(healthcheck.RESTMode),
 		},
@@ -43,7 +57,8 @@ func makeAutoHealthCheckTasks(clusterID uuid.UUID) []*scheduler.Task {
 			Enabled:   true,
 			Name:      "alternator",
 			Sched: scheduler.Schedule{
-				Cron: scheduler.NewCronEvery(15 * time.Second),
+				Cron:     scheduler.NewCronEvery(15 * time.Second),
+				Timezone: localTimezone(),
 			},
 			Properties: healthCheckModeProperties(healthcheck.AlternatorMode),
 		},
@@ -60,6 +75,7 @@ func makeAutoRepairTask(clusterID uuid.UUID) *scheduler.Task {
 		Name:      "all-weekly",
 		Sched: scheduler.Schedule{
 			Cron:       scheduler.MustCron("0 23 * * SAT"),
+			Timezone:   localTimezone(),
 			NumRetries: 3,
 		},
 		Properties: emptyProperties,
