+++ schema/v3.0.0.cql
@@ -1,5 +1,6 @@
 ALTER TYPE schedule ADD cron text;
 ALTER TYPE schedule ADD window list<text>;
+ALTER TYPE schedule ADD timezone text;
 ALTER TYPE schedule ADD retry_wait bigint;
 
 -- CALL rewriteHealthCheck30;
