+++ swagger/scylla-manager.json
@@ -253,6 +253,9 @@
             "type": "string"
           }
         },
+        "timezone": {
+          "type": "string"
+        },
         "start_date": {
           "type": "string",
           "format": "date-time",
